// packages/velix-core/src/types.d.ts
export type RouteType = 'page' | 'api' | 'layout' | 'loading' | 'error' | 'not-found';
export interface Route {
    type: RouteType;
    path: string;
    filePath: string;
    pattern: RegExp;
    segments: string[];
    layout?: string | null;
    loading?: string | null;
    error?: string | null;
    notFound?: string | null;
    template?: string | null;
    middleware?: string | null;
    isServerComponent?: boolean;
    isClientComponent?: boolean;
    isIsland?: boolean;
    params?: Record<string, string>;
}
export interface RouteTree {
    pages: Route[];
    api: Route[];
    layouts: Map<string, string>;
    tree: RouteTreeNode;
    appRoutes: Route[];
    rootLayout?: string;
}
export interface RouteTreeNode {
    children: Record<string, RouteTreeNode>;
    routes: Route[];
}